moved faq page to app
